% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_eto.R
\name{clean_eto}
\alias{clean_eto}
\title{Clean ETO data prior to merging with NWEA data
The data set to be cleaned is an export from the "[LAYCCA] students_roster" report}
\usage{
clean_eto(df)
}
\arguments{
\item{df}{dataframe: NWEA data exported from the website}
}
\value{
dataframe
}
\description{
Clean ETO data prior to merging with NWEA data
The data set to be cleaned is an export from the "[LAYCCA] students_roster" report
}
\examples{
eto <- load_csv(data_folder = './path/to/data_folder')
eto <- clean_eto(df = eto)
}

