% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nwea.R
\name{clean_nwea}
\alias{clean_nwea}
\title{Clean NWEA raw data (as exported from the website) so it can be merged with ETO records}
\usage{
clean_nwea(df)
}
\arguments{
\item{df}{dataframe: NWEA data exported from the website}
}
\value{
dataframe
}
\description{
Clean NWEA raw data (as exported from the website) so it can be merged with ETO records
}
\examples{
nwea <- load_csv(data_folder = './path/to/data_folder')
nwea <- clean_nwea(df = nwea)
}

