% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_nwea_tp.R
\name{clean_nwea_tp}
\alias{clean_nwea_tp}
\title{Clean NWEA data exported from ETO ("[Admin] raw_touchpoint_report_detailed")}
\usage{
clean_nwea_tp(df, expected_columns = c("subject_id", "name", "tp_name",
  "date", "question_id", "question", "question_short", "answer_id", "answer",
  "answer_weight", "program_name", "program_id"),
  kept_columns = c("subject_id", "name", "date", "question_short",
  "answer_id", "answer"), expected_terms = c("fall", "winter", "spring"),
  score_min = 100, score_max = 300)
}
\arguments{
\item{df}{dataframe: NWEA data exported from ETO}
}
\value{
dataframe
}
\description{
Clean NWEA data exported from ETO ("[Admin] raw_touchpoint_report_detailed")
}
\examples{
nwea <- load_txt("./path/to/nwea_scores.txt")
nwea <- clean_nwea_tp(df = nwea)
}

