% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_roster.R
\name{clean_roster}
\alias{clean_roster}
\title{Clean roster data prior to merging with NWEA data
The data set to be cleaned is an export from the "[LAYCCA] students_roster" report}
\usage{
clean_roster(df, expected_columns = c("id", "subject_id", "fname", "lname",
  "dob", "ell", "ethnicity", "race", "start", "end", "days_enrolled"),
  kept_columns = c("id", "subject_id", "fname", "lname", "dob"))
}
\arguments{
\item{df}{dataframe: NWEA data exported from the website}

\item{expected_columns}{character vector: Expected column names for df}

\item{kept_columns}{character vector: Column names from df to keep}
}
\value{
dataframe
}
\description{
Clean roster data prior to merging with NWEA data
The data set to be cleaned is an export from the "[LAYCCA] students_roster" report
}
\examples{
eto <- load_csv(data_folder = './path/to/data_folder')
eto <- clean_roster(df = eto)
}

