% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_change}
\alias{compute_change}
\title{compute_change()}
\usage{
compute_change(df, col = c("subject_id", "prepost", "score"))
}
\arguments{
\item{df}{dataframe: a dataframe returned by id_prepost()}

\item{col}{character vector: name of columns to keep to compute change in scores.}
}
\value{
dataframe
}
\description{
This is a helper function that computes change in score between pre and post test. The data must be pre-processed by the followng function: id_prepost()
}

