% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_nwea.R
\name{export_nwea}
\alias{export_nwea}
\title{export_nwea()}
\usage{
export_nwea(df, path = "./")
}
\arguments{
\item{df}{dataframe: formatted NWEA data}
}
\value{
csv file
}
\description{
Helper function to export formatted NWEA data as .csv for upload in ETO
}
\examples{
nwea <- load_csv(data_folder = './path/to/data_folder')
nwea <- clean_nwea(df = nwea)
nwea <- format_nwea(df = nwea)
export(df = nwea, path = './output')

}

