% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_nwea.R
\name{format_nwea}
\alias{format_nwea}
\title{format_nwea()}
\usage{
format_nwea(df)
}
\arguments{
\item{df}{dataframe: NWEA data exported from the website}
}
\value{
dataframe
}
\description{
Helper function to format NWEA data for upload in ETO
}
\examples{
nwea <- load_csv(data_folder = './path/to/data_folder')
nwea <- clean_nwea(df = nwea)
nwea <- format_nwea(df = nwea)

}

