% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_iqr}
\alias{get_iqr}
\title{get_iqr
Takes a subset of a norms_status data as input. Return a data frame with the necesary
information to create a boxplot.}
\usage{
get_iqr(df)
}
\arguments{
\item{df}{dataframe: Subset of norms_status_xxxx data}
}
\value{
dataframe
}
\description{
get_iqr
Takes a subset of a norms_status data as input. Return a data frame with the necesary
information to create a boxplot.
}
\examples{
library(dplyr)
data(norms_status_2015)
df <- norms_status_2015 \%>\%
 filter(season == 4) \%>\%
 filter(grade == 11) \%>\% 
 filter(subject == 2)

get_iqr(df)

}

