% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{load_nwea}
\alias{load_nwea}
\title{load_nwea()}
\usage{
load_nwea(data_folder)
}
\arguments{
\item{data_folder}{character: path to folder containing NWEA scores .csv files.}
}
\value{
dataframe
}
\description{
Helper function to load and combine various NWEA scores export files
}
\examples{
nwea <- load_nwea(data_folder = './path/to/data_folder')
}

