% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms_2015_raw.R
\docType{data}
\name{norms_2015_raw}
\alias{norms_2015_raw}
\title{NWEA Norms data 2015 (status)}
\format{Data frame with columns
\describe{
\item{season}{Term. Possible Values:
fall,
winter,
spring,
summer}
\item{grade}{Focal grade (the grade regarded as the present grade of focus for the purposes of norms look-up)
Grade 13 refers to kindergarten}
\item{subject}{Measurement Scale.	Possible Values:
1 = Mathematics,
2 = Reading,
3 = Language Usage
4 = General science
As of 2015, Science Concepts and Processes is no longer available}
\item{student_pctile}{Percentile associated with student RIT score for subject, grade, and season Range = 1-99}
\item{rit}{RIT score}
}}
\source{
NWEA, Northwest Evaluation Association,
 \url{https://legacysupport.nwea.org/support/article/norms-study-resources}
}
\usage{
norms_2015_raw
}
\description{
Data from the RIT Scale Norms Study 2015
}
\keyword{datasets}

