% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot_distribution}
\alias{plot_distribution}
\title{plot_distribution()}
\usage{
plot_distribution(df)
}
\arguments{
\item{df}{data.frame: Name of the data frame containing NWEA data}
}
\value{
ggplot
}
\description{
plot distribution of RIT reading scores
}

