% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_change}
\alias{recode_change}
\title{recode_change()}
\usage{
recode_change(vector)
}
\arguments{
\item{numeric}{vector}
}
\value{
factor vector
}
\description{
This is a helper function that recode a numeric vector into 3 categories: positive, no change, negative
}

