% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rit_to_efl.R
\docType{data}
\name{rit_to_efl}
\alias{rit_to_efl}
\title{Equivalency table for NWEA RIT reading, TABE Reading, Lexile scores, and NRS EFL levels}
\format{Data frame with columns
\describe{
\item{rit}{NWEA RIT Reading scores}
\item{tabe_reading}{TABE reading scores}
\item{lexile}{Lexile scores}
\item{nre_efl}{NRS EFL levels}
}}
\source{
Oregon Department of Education
 \url{http://www.ode.state.or.us/teachlearn/subjects/elarts/reading/literacy/summerinstitute/2006/ritlexiletable.pdf}
}
\usage{
rit_to_efl
}
\description{
Data from RIT Lexile table
}
\keyword{datasets}

