% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rit_to_grade}
\alias{rit_to_grade}
\title{rit_to_grade()}
\usage{
rit_to_grade(rit, ref_table)
}
\arguments{
\item{rit}{dataframe: a dataframe containing RIT reading scores to be converted}

\item{ref_table}{dataframe: reference table to be used for the conversion of RIT scores to grades}

\item{col}{character vector: name of columns to keep to compute change in scores.}
}
\value{
dataframe
}
\description{
This is a helper function that converts RIT reading scores into grade equivalent. The conversion is based on the NWEA Normative Data Reference  \url{https://www.nwea.org/resources/2015-normative-data/}
}

