% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norms_status_2015.R
\docType{data}
\name{norms_status_2015}
\alias{norms_status_2015}
\title{NWEA Norms data 2015 (status)}
\format{An object of class \code{data.frame} with 11880 rows and 5 columns.}
\source{
NWEA, Northwest Evaluation Association,
 \url{https://legacysupport.nwea.org/support/article/norms-study-resources}
}
\usage{
norms_status_2015
}
\description{
Data from the RIT Scale Norms Study 2015
}
\keyword{datasets}

