% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_plots.R
\name{plot_individual_scores}
\alias{plot_individual_scores}
\title{plot_individual_scores()}
\usage{
plot_individual_scores(df, my_palette = c("#2CA02C", "#FF7F0E", "#D62728"))
}
\arguments{
\item{df}{data.frame: Name of the data frame containing NWEA data}
}
\value{
ggplot
}
\description{
plot RIT reading score for each students on the roster
}
\examples{
nwea <- load_nwea(data_folder = './path/to/data_folder')
}

